+++ packages/block-editor/src/components/colors/utils.js
@@ -1,7 +1,7 @@
 import { colord, extend } from 'colord';
 import namesPlugin from 'colord/plugins/names';
 import a11yPlugin from 'colord/plugins/a11y';
-import { kebabCase } from '@wordpress/string';
+import { kebabCase } from '@wordpress/kebab-case';
 
 extend( [ namesPlugin, a11yPlugin ] );