+++ packages/dependency-extraction-webpack-plugin/lib/util.js
@@ -7,7 +7,7 @@ const BUNDLED_PACKAGES = [
 	'@wordpress/grid',
 	'@wordpress/icons',
 	'@wordpress/interface',
-	'@wordpress/string',
+	'@wordpress/kebab-case',
 	'@wordpress/style-runtime',
 	'@wordpress/ui',
 	'@wordpress/undo-manager',