+++ packages/block-library/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Internal
 
--   Embed: Use the new `@wordpress/string` package instead of unlocking the `kebabCase` utility from the `@wordpress/components` private APIs ([#81294](https://github.com/WordPress/gutenberg/pull/81294)).
+-   Embed: Use the new `@wordpress/kebab-case` package instead of unlocking the `kebabCase` utility from the `@wordpress/components` private APIs ([#81294](https://github.com/WordPress/gutenberg/pull/81294)).
 
 ### Enhancements