+++ packages/global-styles-ui/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Internal
 
--   Font Library: Use the new `@wordpress/string` package instead of unlocking the `kebabCase` utility from the `@wordpress/components` private APIs ([#81294](https://github.com/WordPress/gutenberg/pull/81294)).
+-   Font Library: Use the new `@wordpress/kebab-case` package instead of unlocking the `kebabCase` utility from the `@wordpress/components` private APIs ([#81294](https://github.com/WordPress/gutenberg/pull/81294)).
 
 ### Enhancements